FE methods added

+++ Scripts/FeatureExtraction/get_CodeMetrics.py
@@ -13,35 +13,36 @@ def get_CodeMetrics(SamplesFolderName,SamplesDirPath = '',DatasetName = ''):
         else:
             SamplesDirPath = './' + str(SamplesDirPath) + '/' 
         
-        OriginalDestinationPath = str(get_Path('OriginalReports',SamplesFolderName)) + '/'
-        EditedDestinationPath = str(get_Path('EditedReports',SamplesFolderName)) + '/'
-        Raw_CodeMetrics_OutDir = str(get_Path('Raw_CodeMetrics',SamplesFolderName)) + '/'
-        CodeMetrics_OutDir  = str(get_Path('CodeMetrics',SamplesFolderName)) + '/'
+        if os.path.exists(SamplesDirPath):
+            OriginalDestinationPath = str(get_Path('OriginalReports',SamplesFolderName)) + '/'
+            EditedDestinationPath = str(get_Path('EditedReports',SamplesFolderName)) + '/'
+            Raw_CodeMetrics_OutDir = str(get_Path('Raw_CodeMetrics',SamplesFolderName)) + '/'
+            CodeMetrics_OutDir  = str(get_Path('CodeMetrics',SamplesFolderName)) + '/'
 
-        generate_MetricsReports(SamplesDirPath,OriginalDestinationPath)
-        prepare_GeneratedMetricsReports(OriginalDestinationPath,EditedDestinationPath)
-        
-        if DatasetName == '':
-            DatasetName = SamplesFolderName
-        
-        
-        metricsDF = parse_MetricsReports(ReportsFolder = EditedDestinationPath)
-        UniqueFilename = generate_UniqueFilename(DatasetName,'Raw_CodeMetrics')
-        metricsDF.to_csv(Raw_CodeMetrics_OutDir + UniqueFilename + '.csv',index=False)
+            generate_MetricsReports(SamplesDirPath,OriginalDestinationPath)
+            prepare_GeneratedMetricsReports(OriginalDestinationPath,EditedDestinationPath)
+            
+            if DatasetName == '':
+                DatasetName = SamplesFolderName
+            
+            
+            metricsDF = parse_MetricsReports(ReportsFolder = EditedDestinationPath)
+            UniqueFilename = generate_UniqueFilename(DatasetName,'Raw_CodeMetrics')
+            metricsDF.to_csv(Raw_CodeMetrics_OutDir + UniqueFilename + '.csv',index=False)
 
-        Raw_CodeMetrics_OutDir = get_Path('Raw_CodeMetrics_OutDir',SamplesFolderName)
-        
-        print('Done! Raw Metrics data is available in: ' + str(os.path.relpath(str(Raw_CodeMetrics_OutDir) + UniqueFilename + '.csv', Path.cwd().parent)))
+            Raw_CodeMetrics_OutDir = get_Path('Raw_CodeMetrics_OutDir',SamplesFolderName)
+            
+            print('Done! Raw Metrics data is available in: ' + str(os.path.relpath(str(Raw_CodeMetrics_OutDir) + UniqueFilename + '.csv', Path.cwd().parent)))
 
-        preProcessed_metricsDF = preprocesse_MetricsData(metricsDF)
-        UniqueFilename = generate_UniqueFilename(DatasetName,'CodeMetrics')
-        preProcessed_metricsDF.to_csv(CodeMetrics_OutDir + UniqueFilename + '.csv' ,index=False)
-        
-        CodeMetrics_OutDir = get_Path('CodeMetrics_OutDir',SamplesFolderName)
-        print('Done! Code Metrics data is available in: ' + str(os.path.relpath(str(CodeMetrics_OutDir) + UniqueFilename + '.csv', Path.cwd().parent)))
-        display(preProcessed_metricsDF)
-        return True
-    
+            preProcessed_metricsDF = preprocesse_MetricsData(metricsDF)
+            UniqueFilename = generate_UniqueFilename(DatasetName,'CodeMetrics')
+            preProcessed_metricsDF.to_csv(CodeMetrics_OutDir + UniqueFilename + '.csv' ,index=False)
+            
+            CodeMetrics_OutDir = get_Path('CodeMetrics_OutDir',SamplesFolderName)
+            print('Done! Code Metrics data is available in: ' + str(os.path.relpath(str(CodeMetrics_OutDir) + UniqueFilename + '.csv', Path.cwd().parent)))
+            display(preProcessed_metricsDF)
+        else:
+            print(f'Samples directory does not exist: {SamplesDirPath}')
     except Exception as err:
         print(f"Unexpected {err=}, {type(err)=}")
         raise
